// src/app/lib/types.ts
export interface selectType {
  id: string | number;
  value: string;
}
// ...
export interface CostCenterType {
  idCostCenter?: number;
  uniqueId?: string;
  costCenter: string;
  description: string;
  bidang: string;
  dinas: string;
  directorat: string;
  groupDinas: string;
  profitCenter: string;
  active: boolean;
  createdAt?: Date;
  createdBy?: string;
  updatedAt?: Date;
  updatedBy?: string | null;
}